:sparkles: 엔티티에 데이터 이관
