+++ src/features/social/ui/SocialItem.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled, { keyframes } from 'styled-components';
-import { Friend } from '../model/types/friend.types';
+import { Friend } from '../../../entities/friend';
 import { commonColors, commonShadows, commonBorderRadius } from '../../../shared/ui/commonStyles';
 
 const fadeInUp = keyframes`
