+++ src/shared/lib/hooks/useAuth.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { login, signup, loginCheckApi, User } from '../../../features/auth/api';
+import { login, signup, loginCheckApi } from '../../../features/auth/api';
 import { useAuth as useAuthContext } from '../context/AuthContext';
 
 export const useAuth = () => {
